requirements.txt
alembic==1.9.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
charset-normalizer==2.1.1
cmaes==0.9.1
colorlog==6.7.0
conda-pack==0.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dgl==1.1.0+cu117
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
gputil==1.4.0
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter-client==7.4.4
jupyter-core==4.11.2
jupyter-server==1.23.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
littleutils==0.2.2
mako==1.2.4
markupsafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.4
ogb==1.3.5
optuna==3.1.0
outdated==0.2.2
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pyyaml==6.0
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
send2trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlalchemy==1.4.46
stack-data==0.6.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
zipp==3.10.0